Common: image_file module, load and save BitmapData, not Bitmap
