+++ Common/CMakeLists.txt
@@ -93,6 +93,8 @@ target_sources(common
     gfx/bitmapdata.cpp
     gfx/bitmapdata.h
     gfx/gfx_def.h
+    gfx/image_file.cpp
+    gfx/image_file.h
     gui/guibutton.cpp
     gui/guibutton.h
     gui/guidefines.h
