+++ Common/gfx/bitmapdata.cpp
@@ -19,6 +19,21 @@ namespace AGS
 namespace Common
 {
 
+uint32_t BitmapData::GetPixel(int x, int y) const
+{
+    const uint8_t *line = GetLine(y);
+    switch (_bitsPerPixel)
+    {
+    case 1: return (line[x / 8] >> (7 - x % 8)) & 0x1;
+    case 4: return (line[x / 2] >> (1 - x % 2)) & 0x4;
+    case 8: return line[x];
+    case 16: return *reinterpret_cast<const uint16_t*>(&line[x * 2]);
+    case 32: return *reinterpret_cast<const uint32_t*>(&line[x * 4]);
+    default: assert(false); return 0;
+    }
+}
+
+
 namespace PixelOp
 {
