+++ Solutions/Common.Lib/Common.Lib.vcxproj.filters
@@ -512,6 +512,9 @@
     <ClCompile Include="..\..\Common\gfx\bitmapdata.cpp">
       <Filter>Source Files\gfx</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\Common\gfx\image_file.cpp">
+      <Filter>Source Files\gfx</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\Common\ac\audiocliptype.h">
@@ -824,6 +827,9 @@
       <Filter>Header Files\debug</Filter>
     </ClInclude>
     <ClInclude Include="..\..\Common\gfx\bitmapdata.h">
+      <Filter>Header Files\gfx</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Common\gfx\image_file.h">
       <Filter>Header Files\gfx</Filter>
     </ClInclude>
   </ItemGroup>
